+++ src/commands/compile.ts
@@ -8,6 +8,7 @@ import { getAnyTTFromFolder } from '../parsers/open';
 import { parseString } from 'xml2js';
 import * as path from 'path';
 import { saveLWC } from './saveLWC';
+import { createPackageXML, deployFiles } from './deploy';
 
 export default function compile(document: vscode.TextDocument): Promise<any> {
     if(!document) {
@@ -48,7 +49,23 @@ export default function compile(document: vscode.TextDocument): Promise<any> {
     }
 
     // Start doing stuff
-    if (folderToolingType && toolingType === undefined) {
+    if (folderToolingType === 'EmailTemplate' || folderToolingType === 'Document') {
+        return createPackageXML([document.fileName], vscode.window.forceCode.storageRoot)
+            .then(() => {
+                const files: string[] = [];
+                var pathSplit = 'documents';
+                if(folderToolingType === 'EmailTemplate') {
+                    pathSplit = 'email';
+                }
+                var foldName: string = document.fileName.split(path.sep + pathSplit + path.sep).pop();
+                //foldName = foldName.substring(0, foldName.lastIndexOf('.'));
+                files.push(path.join(pathSplit, foldName));
+                files.push(path.join(pathSplit, foldName + '-meta.xml'));
+                files.push('package.xml');
+                return deployFiles(files, vscode.window.forceCode.storageRoot);
+            })
+            .catch(onError);
+    } else if (folderToolingType && toolingType === undefined) {
         // This process uses the Metadata API to deploy specific files
         // This is where we extend it to create any kind of metadata
         // Currently only Objects and Permission sets ...