+++ src/commands/saveLWC.ts
@@ -5,50 +5,47 @@ import { FCFile } from '../services/codeCovView';
 import diff from './diff';
 import * as forceCode from './../forceCode';
 import * as path from 'path';
-import { deployFiles, createPackageXML } from './deploy';
+import { createPackageXML, deployFiles } from './deploy';
 
 // =======================================================================================================================================
 // ================================                Lightning Components               ===========================================
 // =======================================================================================================================================
-export function saveAura(document: vscode.TextDocument, toolingType: string, Metadata?: {}): Promise<any> {
+export function saveLWC(document: vscode.TextDocument, toolingType: string, Metadata?: {}): Promise<any> {
+    const fileName: string = document.fileName.split(path.sep).pop();
     const name: string = parsers.getName(document, toolingType);
-    const ext: string = parsers.getFileExtension(document);
-    var DefType: string = getAuraDefTypeFromDocument(document);
-    var Format: string = getAuraFormatFromDocument();
+    const Format: string = parsers.getFileExtension(document);
     var Source: string = document.getText();
     var currentObjectDefinition: any = undefined;
-    // Aura Bundles are a special case, since they can be upserted with the Tooling API
-    // Instead of needing to be compiled, like Classes and Pages..
     if(Metadata) {
         return Promise.resolve(vscode.window.forceCode)
-            .then(getAuraBundle)
-            .then(ensureAuraBundle)
+            .then(getLWCBundle)
+            .then(ensureLWCBundle)
             .then(updateMetaData);
     }
     return Promise.resolve(vscode.window.forceCode)
-        .then(getAuraBundle)
-        .then(ensureAuraBundle)
+        .then(getLWCBundle)
+        .then(ensureLWCBundle)
         .then(bundle => {
-            return getAuraDefinition(bundle)
-                .then(definitions => upsertAuraDefinition(definitions, bundle));
+            return getLWCDefinition(bundle)
+                .then(definitions => upsertLWCDefinition(definitions, bundle));
         });
     
-    function getAuraBundle() {
-        return vscode.window.forceCode.conn.tooling.sobject('AuraDefinitionBundle').find({
+    function getLWCBundle() {
+        return vscode.window.forceCode.conn.tooling.sobject('LightningComponentBundle').find({
             'DeveloperName': name, NamespacePrefix: vscode.window.forceCode.config.prefix || ''
         });
     }
-    function ensureAuraBundle(results) {
+    function ensureLWCBundle(results) {
         // If the Bundle doesn't exist, create it, else Do nothing
         if (results.length === 0 || !results[0]) {
-            // Create Aura Definition Bundle
+            // Create LWC Bundle
             return createPackageXML([document.fileName], vscode.window.forceCode.storageRoot)
                 .then(() => {
                     const files: string[] = [];
-                    files.push(path.join('aura', name));
+                    files.push(path.join('lwc', name));
                     files.push('package.xml');
                     return deployFiles(files, vscode.window.forceCode.storageRoot)
-                        .then(getAuraBundle)
+                        .then(getLWCBundle)
                         .then(bundle => {
                             results[0] = bundle;
                             var newWSMember: forceCode.IWorkspaceMember = {
@@ -58,7 +55,7 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
                                 lastModifiedDate: (new Date()).toISOString(),
                                 lastModifiedByName: '',
                                 lastModifiedById: fcConnection.currentConnection.orgInfo.userId,
-                                type: 'AuraDefinitionBundle',
+                                type: 'LightningComponentBundle',
                                 saveTime: true
                             }
                             codeCovViewService.addClass(newWSMember);
@@ -71,7 +68,7 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
     }
 
     function updateMetaData(bundle) {
-        return vscode.window.forceCode.conn.tooling.sobject('AuraDefinitionBundle').update({
+        return vscode.window.forceCode.conn.tooling.sobject('LightningComponentBundle').update({
             Metadata: Metadata,
             Id: bundle[0].Id
         }).then(res => {
@@ -81,14 +78,14 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
         });
     }
 
-    function getAuraDefinition(bundle) {
-        return vscode.window.forceCode.conn.tooling.sobject('AuraDefinition').find({
-            'AuraDefinitionBundleId': bundle[0].Id
+    function getLWCDefinition(bundle) {
+        return vscode.window.forceCode.conn.tooling.sobject('LightningComponentResource').find({
+            'LightningComponentBundleId': bundle[0].Id
         });
     }
-    function upsertAuraDefinition(definitions, bundle) {
+    function upsertLWCDefinition(definitions, bundle) {
         // If the Definition doesn't exist, create it
-        var def: any[] = definitions.filter(result => result.DefType === DefType);
+        var def: any[] = definitions.filter(result => result.FilePath.split('/').pop() === fileName);
         currentObjectDefinition = def.length > 0 ? def[0] : undefined;
         var curFCFile: FCFile = codeCovViewService.findById(bundle[0].Id ? bundle[0].Id : bundle[0].id);
         if (currentObjectDefinition !== undefined) {
@@ -99,15 +96,15 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
                         return {};
                     }
                     if (s === 'Overwrite') {
-                        return updateAura(curFCFile);
+                        return updateLWC(curFCFile);
                     }
                     return {};
                 });
             } else {
-                return updateAura(curFCFile);
+                return updateLWC(curFCFile);
             }
         } else if (bundle[0]) {
-            return vscode.window.forceCode.conn.tooling.sobject('AuraDefinition').create({ AuraDefinitionBundleId: bundle[0].Id ? bundle[0].Id : bundle[0].id, DefType, Format, Source }).then(res => {
+            return vscode.window.forceCode.conn.tooling.sobject('LightningComponentResource').create({ LightningComponentBundleId: bundle[0].Id ? bundle[0].Id : bundle[0].id, Format, Source }).then(res => {
                 if(curFCFile) {
                     var tempWSMem: forceCode.IWorkspaceMember = curFCFile.getWsMember();
                     tempWSMem.lastModifiedDate = (new Date()).toISOString();
@@ -124,8 +121,8 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
         return undefined;
     }
 
-    function updateAura(fcfile: FCFile) {
-        return vscode.window.forceCode.conn.tooling.sobject('AuraDefinition').update({ Id: currentObjectDefinition.Id, Source }).then(res => {
+    function updateLWC(fcfile: FCFile) {
+        return vscode.window.forceCode.conn.tooling.sobject('LightningComponentResource').update({ Id: currentObjectDefinition.Id, Source }).then(res => {
             if(fcfile) {
                 var tempWSMem: forceCode.IWorkspaceMember = fcfile.getWsMember();
                 tempWSMem.lastModifiedDate = (new Date()).toISOString();
@@ -138,70 +135,4 @@ export function saveAura(document: vscode.TextDocument, toolingType: string, Met
         })
         .catch(err => { return { State: 'Error', message: err.message ? err.message : err }});
     }
-
-    function getAuraFormatFromDocument() {
-        // is 'js', 'css', or 'xml'
-        switch (ext) {
-            case 'js':
-                return 'JS';
-            case 'css':
-                return 'CSS';
-            default:
-                return 'XML';
-        }
-    }
-}
-
-export function getAuraDefTypeFromDocument(document: vscode.TextDocument) {
-    const fname: string = parsers.getName(document, 'AuraDefinition');
-    const extension: string = parsers.getFileExtension(document);
-    const fileName: string = parsers.getFileName(document);
-    switch (extension) {
-        case 'app':
-            // APPLICATION — Lightning Components app
-            return 'APPLICATION';
-        case 'cmp':
-            // COMPONENT — component markup
-            return 'COMPONENT';
-        case 'auradoc':
-            // DOCUMENTATION — documentation markup
-            return 'DOCUMENTATION';
-        case 'css':
-            // STYLE — style (CSS) resource
-            return 'STYLE';
-        case 'evt':
-            // EVENT — event definition
-            return 'EVENT';
-        case 'design':
-            // DESIGN — design definition
-            return 'DESIGN';
-        case 'svg':
-            // SVG — SVG graphic resource
-            return 'SVG';
-        case 'js':
-            var fileNameEndsWith: string = fileName.replace(fname, '').toLowerCase();
-            if (fileNameEndsWith === 'controller') {
-                // CONTROLLER — client-side controller
-                return 'CONTROLLER';
-            } else if (fileNameEndsWith === 'helper') {
-                // HELPER — client-side helper
-                return 'HELPER';
-            } else if (fileNameEndsWith === 'renderer') {
-                // RENDERER — client-side renderer
-                return 'RENDERER';
-            };
-            break;
-        case 'tokens':
-            return 'TOKENS';
-        case 'intf':
-            return 'INTERFACE';
-        case 'xml':
-            return 'Metadata';
-        default:
-            throw `Unknown extension: ${extension} .`;
-    }
-    // Yet to be implemented
-    // PROVIDER — reserved for future use
-    // TESTSUITE — reserved for future use
-    // MODEL — deprecated, do not use
 }